Time is now measured in milliseconds instead of nanoseconds, added Timer class, tank respawn and shield animations, the game window is now 3 times bigger at the start than before

+++ src/Game/GameObjects/BetonWall.h
@@ -39,7 +39,7 @@ public:
 
     BetonWall(const EBetonWallType eBetonWallType, const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void render() const override;
-    virtual void update(const uint64_t delta);
+    virtual void update(const double delta);
 
 private:
     std::array<bool, 4> m_currentBlockState;
