+++ src/Game/GameObjects/BrickWall.cpp
@@ -88,5 +88,5 @@ void BrickWall::render() const
     renderBrick(EBrickLocation::BottomRight);
 }
 
-void BrickWall::update(const uint64_t delta)
+void BrickWall::update(const double delta)
 {}
