+++ src/Game/GameObjects/BrickWall.h
@@ -59,7 +59,7 @@ public:
 
     BrickWall(const EBrickWallType eBrickWallType, const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void render() const override;
-    virtual void update(const uint64_t delta);
+    virtual void update(const double delta);
 
 private:
     void renderBrick(const EBrickLocation eBrickLocation) const;
