+++ src/Renderer/SpriteAnimator.cpp
@@ -10,7 +10,7 @@ namespace RenderEngine
     , m_currentAnimationTime(0)
     {}
 
-    void SpriteAnimator::update(const uint64_t delta)
+    void SpriteAnimator::update(const double delta)
     {
         m_currentAnimationTime += delta;
